+++ app/src/main/java/dev/tuandoan/tasktracker/data/database/TagInfo.kt
@@ -1,0 +1,3 @@
+package dev.tuandoan.tasktracker.data.database
+
+data class TagInfo(val tag: String, val tagColor: String?, val taskCount: Int)